06_Automation/post_scheduler.py
#!/usr/bin/env python3
"""
Archive 35 — Post Scheduler
Reads schedule CSV and publishes posts at scheduled times.

Status: PLACEHOLDER — Not yet implemented
"""

import os
import sys
from datetime import datetime
from pathlib import Path

import pandas as pd
import yaml
# import schedule  # Uncomment when implementing


def load_config(config_path: str = "config.yaml") -> dict:
    """Load configuration from YAML file."""
    with open(config_path, "r") as f:
        return yaml.safe_load(f)


def load_schedule(schedule_path: str) -> pd.DataFrame:
    """Load posting schedule from CSV."""
    return pd.read_csv(schedule_path)


def get_pending_posts(df: pd.DataFrame) -> pd.DataFrame:
    """Filter for posts that are due and not yet posted."""
    now = datetime.now()
    # TODO: Implement filtering logic
    return df[df["status"] == "scheduled"]


def post_to_instagram(image_path: str, caption: str) -> bool:
    """Post image to Instagram."""
    # TODO: Implement using instagrapi
    print(f"[PLACEHOLDER] Would post to Instagram: {image_path}")
    return False


def post_to_facebook(image_path: str, caption: str) -> bool:
    """Post image to Facebook."""
    # TODO: Implement using facebook-sdk
    print(f"[PLACEHOLDER] Would post to Facebook: {image_path}")
    return False


def post_to_tiktok(video_path: str, caption: str) -> bool:
    """Post video to TikTok."""
    # TODO: Implement using TikTokApi
    print(f"[PLACEHOLDER] Would post to TikTok: {video_path}")
    return False


def main():
    """Main scheduler loop."""
    print("Archive 35 Post Scheduler")
    print("=" * 40)
    print("Status: PLACEHOLDER — Not yet implemented")
    print()
    print("To implement:")
    print("1. Set up API credentials in config.yaml")
    print("2. Implement platform posting functions")
    print("3. Set up scheduling logic")
    print("4. Run as daemon or cron job")


if __name__ == "__main__":
    main()
